Fix the release call: vmDeallocateFn takes UInt, not UInt64
VmDeallocateFn is declared (MachPort, UInt, MachVmSize), so passing m.localBase straight in is a type error - the local tools here cannot catch that, only the compiler could, and it did. localBase is a UInt64 address; UInt is 64-bit on arm64 so the conversion is width-preserving. Also re-checked every other @convention(c) call site I touched: readSmart's UInt64(n) <= rec.remain and UInt(rec.localBase), mappedRecord's internal arithmetic, and the readBytesDirect free which already passed UInt and MachVmSize.

+++ Aether/FangUI/Memory/MemoryProbe.swift
@@ -194,7 +194,8 @@ final class MemoryProbe {
     private static func releaseAllMappings() {
         if let fn = vmDeallocateFn {
             for m in mappedRanges {
-                _ = fn(mach_task_self_, m.localBase, m.size)
+                // 第二个形参是 UInt（mach_vm_address_t），我们的 localBase 是 UInt64 —— arm64 上同宽
+                _ = fn(mach_task_self_, UInt(m.localBase), m.size)
             }
         }
         mappedRanges.removeAll()
